Fixing Movie fetching.
The initial search would find a movie in the OpenSubtitles API, and, if it wasn't in the DB yet, download subtitles and save.
This was then replaced by the Full-Text engine that you can use today. This old search still exists but is not pretty at the moment :blush:. You can find it in /movies.